Add custom exception handler and middleware

+++ Restaurant/Business/Services/DishService.cs
@@ -49,15 +49,15 @@ namespace Business.Services
         {
             if (model == null)
             {
-                throw new RestaurantException("Model is null");
+                throw new RestaurantException("Dish model is null");
             }
-            if (model.Price < 0)
-            {
-                throw new RestaurantException("Price is negative");
+            if (model.Price <= 0)
+            { 
+                throw new RestaurantException("Dish price is <= 0");
             }
             if (String.IsNullOrEmpty(model.Name))
             {
-                throw new RestaurantException("Name is empty");
+                throw new RestaurantException("Dish name is empty");
             }
                      
 
@@ -71,9 +71,16 @@ namespace Business.Services
 
         public async Task DeleteAsync(int modelId)
         {
-            await _unitOfWork.DishRepository.DeleteByIdAsync(modelId);
+            var dish = await _unitOfWork.DishRepository.GetByIdWithDetailsAsync(modelId);
 
-            await _unitOfWork.SaveAsync();
+            if (dish != null)
+            {
+                await _unitOfWork.DishRepository.DeleteByIdAsync(modelId);
+
+                await _unitOfWork.SaveAsync();
+            }
+
+
         }
 
         public async Task<IEnumerable<DishModel>> GetAllAsync()
@@ -119,19 +126,20 @@ namespace Business.Services
         }
 
         public async Task UpdateAsync(DishModel model)
-        {
+        {            
+
             if (model == null)
             {
-                throw new RestaurantException("Model is null");
+                throw new RestaurantException("Dish Model is null");
             }
             if (String.IsNullOrEmpty(model.Name))
             {
-                throw new RestaurantException("Name is empty");
+                throw new RestaurantException("Dish Name is empty");
             }
 
-            if (model.Price < 0)
+            if (model.Price <= 0)
             {
-                throw new RestaurantException("Price is negative");
+                throw new RestaurantException("Dish Price is <=0 ");
             }
 
             var dish = _mapper.Map<Dish>(model);
